Forgot that my cg-makea.chpl test relied on changes to the Random module which I'd made but never checked in, resulting in last night's regression.
This commit makes those changes, adding a getNext()
function to the RandomStream class so that values
can be plucked out one at a time as in CG's
initialization.  In doing so, it cleans up the internal
state of the RandomStream class a bit, replacing the
numGenerated count with a cursorVal cursor to track
where in the random stream we currently are.  This
results in equivalent behavior as well as being
incrementally faster.

In cleaning up Random.chpl for this commit, had to
modify the way I used it in cg-makea.chpl a bit, which
exposed a bug in my thinking and modified the output
a bit.  Still haven't verified it against that which
is generated by the official NPB version.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9454 3a8e244f-b0f2-452b-bcba-4c88e055c3ca