Next pass at cg-makea.chpl -- the iterator is now returning the correct pattern of sparse indices and values for the sparse matrix in CG.  In order to get this to work, I had to rework the Random module so that it didn't round odd seeds up to the next even seed value (the NPB doesn't do this and relies on an odd seed).  Working with MB to try and remember why we built that into the code.
Turns out that our HPCC benchmark codes also all used
odd seeds, so changed these to the next larger even
seed in the studies/hpcc directory (to preserve the
output), and changed the output of the tests in the
release/ subdirectory (to keep the codes looking the
same to the end-user).

In doing this, also cleaned up the Random module a bit to
make it use Chapel better as well as the fact that the
randlc functions were embedded in a class and could use
that to store state.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9466 3a8e244f-b0f2-452b-bcba-4c88e055c3ca