Changed --no-gc to --ref-count turning reference counting off by default and enabled with this flag.
Updated the bane of my regression testing existence
  compflags/bradc/help/userhelp.chpl.

Added a few compopts files to the testing system to use --ref-count on
tests that require it thus continuing the testing of this
functionality.

Updated a few memory tests since classes by default no longer have a
reference counter and lock field.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9478 3a8e244f-b0f2-452b-bcba-4c88e055c3ca