Initial (bad) implementation of sparse domain/array. Supports addition of indices and indexing of array. Currently can't hold more than 30 nonzero values.
Uses a (naive) tuple of sequences to store the indices
in something approximating unsorted coordinate format.
The array currently has its data stored using a 30
element array, hence the limitation.

Made it so that sparse arrays currently have to
be subdomains of a parent domain that specifies
their bounding boxes (in the parser).

Added a test that creates a sparse domain to describe
a diagonal array, initializes it, and prints out the
array (manually).

Next stop: try to implement coordinate storage or
CSR more realistically.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9560 3a8e244f-b0f2-452b-bcba-4c88e055c3ca