Haven't fully tested, but this change works on the sparse/ and array/ test directories, and I couldn't resist being the only person to check in developer code on a weekday for the first time in who knows how long...
This change adds a notion of out-of-boundness
to sparse arrays such that if a sparse array
is accessed out-of-bounds an error occurs (if
bounds checking is turned on); if it is accessed
in-bounds and at a represented element the array
element is returned; and otherwise the IRV or
"zero" value is returned.  Added two tests to
check this behavior.

In doing this change, added a member?() check
to the SingleLocaleArithmetic domain and to
the BaseDomain class so that the SparseDomain
could store a pointer to its parent (for
whom we don't want to assume any particular
implementation/distribution) and query whether
something was in or out of bounds for it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9573 3a8e244f-b0f2-452b-bcba-4c88e055c3ca