Caused runtime error messages to print out internal module filenames/locations when the --devel flag is turned on.  Otherwise, the user filename/line number is printed as before.
Also, added a --no-devel flag to support the ability
to override a developer environment variable without
unsetting it (decided not to make --devel a toggle
for fear that it would result in confusing behavior).

Changed the --help message to not print out the
compiler version information in order to make
the userhelp.chpl test work in a more traditional
way where the userhelp.good file should be updated
rather than the userhelp.txt file.

Turned out to be troublesome to have two flags (--devel
and --no-devel) change the same variable with our
current argument-parsing scheme, so created a new
flag type "N" (for "No-" flag) which is similar to the
"F" flag except that it supports both positive and
negative variations.  Thus, creating "devel" as an "N"
flag causes "--devel" to work like an "F" flag and
"--no-devel" to work like the "f" version of the flag.
(Could imagine adding an "n" flag that reverses this
sense, much as "F" and "f" are opposites, but didn't
do this now).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9616 3a8e244f-b0f2-452b-bcba-4c88e055c3ca