Added support for a --gdb flag on the compiler-generated executable, much as we support for the compiler itself. Obviously, Steve's various "view" functions aren't supported here since they don't make sense, but otherwise things should be more or less identical to how they work within the compiler.  Please let me know if you find any obvious bugs.  The point of this (as with the compiler flag) is to support the ability to quickly start debugging a particular run without having to explicitly specify a bunch of command-line flags within the gdb session.
Note that while Steve's view commands don't make sense
within this context, we could generate some other commands
that would be very nice, such as mapping functions from
Chapel identifier names to C names and vice-versa, as
a means of helping the user work with our generated
code a bit easier.

In making this change, created a single _chpl_exit()
exit point for Chapel programs to (a) cleanup state
on the way out of the program, and (b) provide a
location to set a gdb breakpoint in case of non-zero
exit.  Also made errors return a non-zero exit code,
which seems appropriate.

Due to the above change, made sub_test not complain if
a test returns a non-zero exit code.  I hope this is
reasonable...  I couldn't think of why this would be
a concern.

Added a test of using the --gdb flag on the generated
executable, similar to the one we have for the compiler
--gdb flag.

Also, adding in some missing .cvsignore files in the
test directories.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9618 3a8e244f-b0f2-452b-bcba-4c88e055c3ca