Improved the default sparse domain/array format to use CRS (Compressed Row Storage).  This removes the 30 nonzeroes limitation from the previous version.  However, there are still some limitations and workarounds:  (1) I don't query the parent domain's row range as I should, but instead hardcode it to be -1..2000, implying that (a) 2002 rows are represented for every sparse domain, and (b) if a row outside this range is specified, an out-of-bounds error will occur;  (2) I'm currently not reallocating the sparse arrays for a domain when a new index is added to it properly.  This means that if the domain is created before the sparse array's elements are defined, things will work; if the sparse array's elements are defined and then the domain is modified they won't.
Added a bunch of tests showing that this format can be used
to express the sparse domain and array from NAS CG, as well
as some simple standalone tests.  Also added some futures
showing the need to reallocate arrays more intelligently.

Fixed a problem with my last checkin to sub_test, in which
a timeout would show up as two errors rather than one, due
to the fact that I forgot to break out of the loop upon
timing out.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9625 3a8e244f-b0f2-452b-bcba-4c88e055c3ca