Change the way we munge overloaded names in the generated code from:
     _<id>_<chapel_name>

to:

     _<chapel_name>_<id>

which makes it easier to view values when debugging the
generated code in gdb.  In particular, you can do:

          (gdb) p myvar

and if myvar isn't found (implying it's been renamed),
you can do:

          (gdb) p _myvar_<TAB>

and TAB completion will complete the variable name if
there's only one visible in the current scope; otherwise,
it will list the candidates (and you will need to inspect
the generated code to determine which you want).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9631 3a8e244f-b0f2-452b-bcba-4c88e055c3ca