Fixed a bug in the sparse implementation in which an off-by-one error could cause extra (redundant) indices from being stored in the sparse domain.
Added a test that counts the number of indices
in a sparse domain to prevent against this kind
of error slipping in unnoticed in the future.

Made the sparse implementation a bit more
efficient by searching a row for a particular
index from top-to-bottom rather than bottom-to-top
to optimize for the case when indices are being
added in sorted order.

Should consider doing a binary rather than linear
search to improve this further.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9651 3a8e244f-b0f2-452b-bcba-4c88e055c3ca