Changed the boundedType field of the range type to an enumerated type called BoundedRangeType with values bounded, boundedLow, and boundedHigh.
Located the BoundedRangeType in the base module to support the
--nostdincs flag.

It appears that a parameter integer cannot be coerced to a parameter
enumerated type so

  range(int,0)

is no longer valid.  It must be

  range(int,bounded)

Updated tests.

Note.  Enumerated types in Java define the constant symbols in the
scope of the enumeration type.  This has some advantages as far as the
name space is concerned.  To access a constant, the programmer writes:

  BoundedRangeType.bounded

Without making this change, we could also take the approach of nesting
BoundedRangeType in the range type and defining some (static?)
constants with the names that we want the user to access.  For
example, they could then specify range.bounded instead of bounded.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9657 3a8e244f-b0f2-452b-bcba-4c88e055c3ca