Fixed a bug in iterators in which the iterator class would not be updated if a field of a record was changed possibly via a method.
This change simplifies the logic in building iterators by storing back
all fields into the iterator class regardless of whether the field was
changed or not.  This fix devastates performance resulting in a
slowdown of 48% on RA at size 2**20.  Improvement is possible, but
correctness trumps performance.

Added a test case to guard this bug.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9764 3a8e244f-b0f2-452b-bcba-4c88e055c3ca