Removed support for the structural interface methods getHeadCursor, getNextCursor, isValidCursor?, and getValue that allow a class or record to be iterated over.  Now use the default iterator named, currently though I'm not especially tied to it, "ault".  This is as in "def ault".
Reworked all structural interface methods to be specified as an ault
iterator.

Added an internal _getIterator function that returns the ault iterator
or an iterator class itself.

Note that tuples and zipper iteration is implemented explicitly via
the structural interface methods currently.

Removed a test that makes use of this structural interface.  Updated
another that illustrates our internal iterator transform.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9778 3a8e244f-b0f2-452b-bcba-4c88e055c3ca