Adding two C versions of the STREAM Triad benchmark. - stream-index.c uses C indexing into the 1D arrays. - stream-walker.c uses a pointer to wak through the   arrays
IMPORTANT NOTE: Currently neither of the vectors are
set up using the same random numbers as in the Chapel
benchmark, they are merely initialized to constant
values.  This should not affect performance, but
should probably be done for consistency.  I knew I
wasn't going to get that done in this sitting, so am
checking in what I have.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9800 3a8e244f-b0f2-452b-bcba-4c88e055c3ca