Enabled forall statements and loops using the bracket syntax to not require an index as in
  A = [Mat] 2.0;

Removed a .future.

Still to do, it does not make sense to require an index in loops that
don't use the special bracket syntax.

There is an ambiguity between an expression statement that is a domain
literal expression and a forall statement using the bracket syntax
without an index followed by an empty statement.  For example, the
code

  [i in D] ;

iterates over D and executes nothing (an empty statement) but the code

  [D] ;

is interpreted as an expression statement that is a domain expression.
There is no iteration over D in this case even though it "could" be
interpreted also as a forall statement over D that executes the empty
statement.  That is, priority is given to the domain expression.

If this ambiguity is concerning, we can remove support for eliding the
iterator from a loop.  I don't think this ambiguity is a concern.  It
should be documented in the spec.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9883 3a8e244f-b0f2-452b-bcba-4c88e055c3ca