Some test additions and changes related to my last check-in.
Removed future for test_sparse1.chpl which now works due to
the support for printing.

Added a test to my CG directory to print out the sparse
domain and array computed by cg-makea.chpl

Added a future to mirror the one I checked in yesterday
afternoon, in which a variable is declared of inferred
sparse domain/array type.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9905 3a8e244f-b0f2-452b-bcba-4c88e055c3ca